pyproject.toml
[project]
name = "2026-225-vochill-cashflow"
version = "0.1.0"
description = "VoChill Cash Flow Forecasting and Reporting System"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pyyaml>=6.0.1",
    "openpyxl>=3.1.2",
    "xlsxwriter>=3.2.0",
    "python-dateutil>=2.8.2",
    "google-cloud-bigquery>=3.17.0",
    "google-auth>=2.27.0",
    "python-dotenv>=1.0.0",
    "db-dtypes>=1.2.0",
]
